#8748 Refactor: Moved ADOJoins to common, subclasses to domain packages
