+++ sormas-backend/src/main/java/de/symeda/sormas/backend/caze/CaseService.java
@@ -150,7 +150,6 @@ import de.symeda.sormas.backend.util.QueryHelper;
 import de.symeda.sormas.backend.visit.Visit;
 import de.symeda.sormas.backend.visit.VisitFacadeEjb;
 import de.symeda.sormas.backend.visit.VisitService;
-import de.symeda.sormas.utils.CaseJoins;
 
 @Stateless
 @LocalBean
