+++ sormas-backend/src/main/java/de/symeda/sormas/backend/event/EventQueryContext.java
@@ -6,7 +6,6 @@ import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.From;
 
 import de.symeda.sormas.backend.common.QueryContext;
-import de.symeda.sormas.utils.EventJoins;
 
 public class EventQueryContext<T> extends QueryContext<T, Event> {
