+++ sormas-backend/src/main/java/de/symeda/sormas/backend/immunization/joins/BaseImmunizationJoins.java
@@ -4,6 +4,7 @@ import javax.persistence.criteria.From;
 import javax.persistence.criteria.Join;
 import javax.persistence.criteria.JoinType;
 
+import de.symeda.sormas.backend.common.AbstractDomainObjectJoins;
 import de.symeda.sormas.backend.immunization.entity.BaseImmunization;
 import de.symeda.sormas.backend.immunization.entity.Immunization;
 import de.symeda.sormas.backend.infrastructure.community.Community;
@@ -11,7 +12,6 @@ import de.symeda.sormas.backend.infrastructure.district.District;
 import de.symeda.sormas.backend.infrastructure.facility.Facility;
 import de.symeda.sormas.backend.infrastructure.region.Region;
 import de.symeda.sormas.backend.person.Person;
-import de.symeda.sormas.backend.util.AbstractDomainObjectJoins;
 
 public class BaseImmunizationJoins<T, S extends BaseImmunization> extends AbstractDomainObjectJoins<T, S> {
