+++ sormas-backend/src/main/java/de/symeda/sormas/backend/location/LocationJoins.java
@@ -20,10 +20,10 @@ import javax.persistence.criteria.Join;
 import javax.persistence.criteria.JoinType;
 
 import de.symeda.sormas.backend.infrastructure.facility.Facility;
+import de.symeda.sormas.backend.common.AbstractDomainObjectJoins;
 import de.symeda.sormas.backend.infrastructure.community.Community;
 import de.symeda.sormas.backend.infrastructure.district.District;
 import de.symeda.sormas.backend.infrastructure.region.Region;
-import de.symeda.sormas.backend.util.AbstractDomainObjectJoins;
 
 public class LocationJoins<T> extends AbstractDomainObjectJoins<T, Location> {
